ixchio-backend/core/cache.py
"""
Semantic cache — saves us 30-50% on API costs.
Uses FAISS for fast L2 similarity + a dict for result storage.
Thread-safe with asyncio.Lock, model loaded lazily on first call.
"""

import asyncio
import numpy as np
import faiss
from sentence_transformers import SentenceTransformer


class SemanticCache:
    def __init__(self, similarity_threshold: float = 0.85):
        self.dimension = 384
        self.index = faiss.IndexFlatL2(self.dimension)
        self.cache = {}
        self.threshold = similarity_threshold
        self.hit_count = 0
        self.miss_count = 0
        self._lock = asyncio.Lock()
        self._counter = 0
        self._model = None

    def _get_model(self):
        if self._model is None:
            self._model = SentenceTransformer("all-MiniLM-L6-v2")
        return self._model

    async def get_or_compute(self, query: str, compute_fn, *args, **kwargs):
        loop = asyncio.get_event_loop()
        model = self._get_model()

        raw = np.ascontiguousarray(
            model.encode([query], normalize_embeddings=True).astype("float32")
        )

        async with self._lock:
            if self.index.ntotal > 0:
                distances, indices = await loop.run_in_executor(
                    None, self.index.search, raw, 1
                )
                similarity = 1 - (distances[0][0] / 2)
                if similarity > self.threshold:
                    self.hit_count += 1
                    return self.cache[indices[0][0]], "cache_hit"

        # miss — actually compute it
        self.miss_count += 1
        result = await compute_fn(*args, **kwargs)

        async with self._lock:
            await loop.run_in_executor(None, self.index.add, raw)
            self.cache[self._counter] = result
            self._counter += 1

        return result, "cache_miss"

    def get_stats(self) -> dict:
        total = self.hit_count + self.miss_count
        return {
            "hits": self.hit_count,
            "misses": self.miss_count,
            "hit_rate": f"{(self.hit_count / total * 100):.1f}%" if total else "0%",
        }
