ixchio-backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.5.0
numpy>=1.24.0,<2.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
sentence-transformers>=2.2.0
tenacity>=8.2.0
rank-bm25>=0.2.2
langgraph>=0.0.40
PyJWT>=2.8.0
pinecone>=6.0.0
